made spinner self-managed & renamed `spinner.tsx` --> `loading.tsx`
